Update to API version 1.2

+++ capture.go
@@ -24,8 +24,8 @@ type Capture struct {
 
 // CaptureParams is a set of params for creating entity.
 type CaptureParams struct {
-	ReconciliationID string  `json:"reconciliation_id,omitempty"`
-	Amount           float64 `json:"amount,omitempty"`
+	ReconciliationID string `json:"reconciliation_id,omitempty"`
+	Amount           int64  `json:"amount,omitempty"`
 }
 
 // New creates new Capture entity.
